[chore] UIColor 확장해서 사용한 코드를 파일 분리함 [refactor] RGB값으로 만든 정적 상수 컬러(레드,그린,그레이) 추가하여 재적용함
