+++ iOS/SignupApp/SignupApp/Views/TextFields/CustomTextField.swift
@@ -8,27 +8,6 @@
 
 import UIKit
 
-extension UIColor {
-    convenience init?(red: Int, green: Int, blue: Int) {
-        guard red >= 0 && red <= 255,
-            green >= 0 && green <= 255,
-            blue >= 0 && blue <= 255
-            else { return nil }
-        self.init(red: CGFloat(red) / 255.0,
-                  green: CGFloat(green) / 255.0,
-                  blue: CGFloat(blue) / 255.0,
-                  alpha: 1.0)
-    }
-    
-    convenience init?(rgb: Int) {
-        self.init(
-            red: (rgb >> 16) & 0xFF,
-            green: (rgb >> 8) & 0xFF,
-            blue: rgb & 0xFF
-        )
-    }
-}
-
 class CustomTextField: UITextField {
     private let messageLabel = MessageLabel()
     
@@ -49,8 +28,7 @@ class CustomTextField: UITextField {
     
     private func setBorder() {
         layer.borderWidth = 0.9
-        let greyColor = UIColor(rgb: 0x707070)!
-        layer.borderColor = greyColor.cgColor
+        layer.borderColor = UIColor.greyColor.cgColor
     }
     
     private func setMessageLabel() {
@@ -82,7 +60,7 @@ extension CustomTextField {
     func setMessageRequireText() {
         let messageRequireText = "필수 항목입니다."
         messageLabel.text = messageRequireText
-        messageLabel.textColor = .red
+        messageLabel.textColor = UIColor.redColor
         messageLabel.isHidden = false
     }
 }
